common/http/code/chapter8/requirements.txt
Flask==2.3.2
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.3